Rename to BlockscoutV2Client
