+++ packages/shared/src/services/blockscout/index.ts
@@ -1,1 +1,3 @@
 export * from './BlockscoutClient'
+export * from './BlockscoutV2Client'
+export * from './model'
